+++ app/src/main/java/joshuatee/wx/settings/SettingsColorPaletteEditor.kt
@@ -32,11 +32,9 @@ import android.util.TypedValue
 import android.view.MenuItem
 import android.widget.EditText
 import androidx.appcompat.widget.Toolbar.OnMenuItemClickListener
-
 import joshuatee.wx.MyApplication
 import joshuatee.wx.R
 import joshuatee.wx.UIPreferences
-import joshuatee.wx.objects.ObjectIntent
 import joshuatee.wx.radar.WXGLNexrad
 import joshuatee.wx.radarcolorpalettes.UtilityColorPalette
 import joshuatee.wx.ui.BaseActivity
@@ -105,7 +103,9 @@ class SettingsColorPaletteEditor : BaseActivity(), OnMenuItemClickListener {
             ObjectDialogue(this, errorCheck)
         }
         val fileName = "colormap" + type + palTitle.text.toString()
-        if (UtilityFileManagement.internalFileExist(context, fileName)) UtilityFileManagement.deleteFile(context, fileName)
+        if (UtilityFileManagement.internalFileExist(context, fileName)) {
+            UtilityFileManagement.deleteFile(context, fileName)
+        }
     }
 
     private fun checkMapForErrors(): String {
@@ -143,7 +143,9 @@ class SettingsColorPaletteEditor : BaseActivity(), OnMenuItemClickListener {
                 }
             }
         }
-        if (lineCount < 2) errors += "Not enough lines present."
+        if (lineCount < 2) {
+            errors += "Not enough lines present."
+        }
         return errors
     }
 
@@ -154,8 +156,6 @@ class SettingsColorPaletteEditor : BaseActivity(), OnMenuItemClickListener {
             R.id.action_help -> ObjectDialogue(this,"Not implemented yet.")
             R.id.action_share -> UtilityShare.textAsAttachment(this, this, palTitle.text.toString(), palContent.text.toString(), "wX_colormap_" + palTitle.text.toString() + ".txt")
             R.id.action_load -> loadSettings()
-            R.id.action_website -> ObjectIntent.showWeb(this, "http://almanydesigns.com/grx/reflectivity/")
-            R.id.action_website2 -> ObjectIntent.showWeb(this, "http://www.usawx.com/grradarexamples.htm")
             else -> return super.onOptionsItemSelected(item)
         }
         return true
@@ -170,9 +170,13 @@ class SettingsColorPaletteEditor : BaseActivity(), OnMenuItemClickListener {
         toolbarBottom.menu.findItem(R.id.action_load).isVisible = true
     }
 
-    private fun loadSettings() { performFileSearch() }
+    private fun loadSettings() {
+        performFileSearch()
+    }
 
-    private fun displaySettings(txt: String) { palContent.setText(txt) }
+    private fun displaySettings(txt: String) {
+        palContent.setText(txt)
+    }
 
     private fun convertPalette(txt: String): String {
         var txtLocal = Utility.fromHtml(txt)